+++ Lab Assignment-1/lab1_pthread.h
@@ -1,8 +1,8 @@
 #ifndef LAB1_PTHREAD_H
 #define LAB1_PTHREAD_H
 
+// #include <malloc.h>
 #include <stdlib.h>
-
 /*
 	Arguments:
 
@@ -39,7 +39,7 @@ void kmeans_pthread(int num_threads,
 					int K,
 					int* data_points,
 					int** data_point_cluster,
-					int** centroids,
+					float** centroids,
 					int* num_iterations
 					);
